requirements.txt
psycopg2-binary
Django~=4.0.3
django-phonenumber-field[phonenumbers]
djangorestframework
pygments